+++ app/lib/frontend/handlers/custom_api.dart
@@ -19,8 +19,8 @@ import '../../scorecard/backend.dart';
 import '../../search/backend.dart';
 import '../../search/search_client.dart';
 import '../../search/search_service.dart';
+import '../../service/topics/count_topics.dart';
 import '../../shared/configuration.dart';
-import '../../shared/count_topics.dart';
 import '../../shared/exceptions.dart';
 import '../../shared/handlers.dart';
 import '../../shared/redis_cache.dart' show cache;
