+++ app/lib/frontend/handlers/misc.dart
@@ -17,9 +17,9 @@ import 'package:shelf/shelf.dart' as shelf;
 import '../../package/backend.dart';
 import '../../package/name_tracker.dart';
 import '../../publisher/backend.dart';
+import '../../service/topics/count_topics.dart';
 import '../../shared/configuration.dart';
 import '../../shared/cookie_utils.dart';
-import '../../shared/count_topics.dart';
 import '../../shared/handlers.dart';
 import '../../shared/redis_cache.dart';
 import '../../shared/urls.dart' as urls;
