+++ app/lib/tool/neat_task/pub_dev_tasks.dart
@@ -17,8 +17,8 @@ import '../../package/backend.dart';
 import '../../search/backend.dart';
 import '../../service/email/backend.dart';
 import '../../service/security_advisories/sync_security_advisories.dart';
+import '../../service/topics/count_topics.dart';
 import '../../shared/configuration.dart';
-import '../../shared/count_topics.dart';
 import '../../shared/datastore.dart';
 import '../../shared/integrity.dart';
 import '../../task/backend.dart';
